requirements.txt
asgiref==3.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.1
uritemplate==4.1.1
